+++ com.project.examsystem/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>com.project.examsystem</name>
+	<name>com.project.student</name>
 	<comment></comment>
 	<projects>
 	</projects>
