+++ com.project.examsystem/src/main/java/com/project/student/dbs/Database.java
@@ -1,0 +1,14 @@
+package com.project.student.dbs;
+
+import java.util.List;
+
+public interface Database<T, I> {
+
+	public List<T> getAll();
+
+	public T findById(I id);
+
+	public void add(T element);
+
+	
+}
