+++ com.project.examsystem/src/main/java/com/project/student/dbs/StudentDatabase.java
@@ -1,0 +1,8 @@
+package com.project.student.dbs;
+
+import com.project.student.models.Student;
+
+public interface StudentDatabase extends Database<Student, Integer> {
+	
+	public String getEmailById(int id);
+}
