+++ com.project.examsystem/src/main/java/com/project/student/models/Student_.java
@@ -1,0 +1,12 @@
+package com.project.student.models;
+
+import javax.persistence.metamodel.SingularAttribute;
+import javax.persistence.metamodel.StaticMetamodel;
+
+@StaticMetamodel(Student.class)
+public abstract class Student_ {
+
+	public static volatile SingularAttribute<Student, Integer> id;
+	public static volatile SingularAttribute<Student, String> email;
+	
+}
