--- com.project.examsystem/src/main/java/it/fi/mirelli/com/project/examsystem/dbs/StudentDatabase.java
@@ -1,8 +1,0 @@
-package it.fi.mirelli.com.project.examsystem.dbs;
-
-import it.fi.mirelli.com.project.examsystem.models.Student;
-
-public interface StudentDatabase extends Database<Student, Integer> {
-	
-	public String getEmailById(int id);
-}
