--- com.project.examsystem/src/main/java/it/fi/mirelli/com/project/examsystem/models/Student_.java
@@ -1,12 +1,0 @@
-package it.fi.mirelli.com.project.examsystem.models;
-
-import javax.persistence.metamodel.SingularAttribute;
-import javax.persistence.metamodel.StaticMetamodel;
-
-@StaticMetamodel(Student.class)
-public abstract class Student_ {
-
-	public static volatile SingularAttribute<Student, Integer> id;
-	public static volatile SingularAttribute<Student, String> email;
-	
-}
