+++ com.project.examsystem/src/test/java/com/project/student/it/EndToEndIT.java
@@ -1,0 +1,15 @@
+package com.project.student.it;
+
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.Persistence;
+
+public class EndToEndIT extends AbstractStudentController {
+
+	@Override
+	protected EntityManagerFactory createEntityManagerFactory() {
+		EntityManagerFactory emf = Persistence.
+				createEntityManagerFactory("examSystem-remote-PU");
+		return emf;
+	}
+
+}
