+++ com.project.examsystem/src/test/java/com/project/student/models/StudentTest.java
@@ -1,0 +1,25 @@
+package com.project.student.models;
+
+import static org.junit.Assert.*;
+
+import org.junit.Before;
+import org.junit.Test;
+
+import com.project.student.models.Student;
+
+public class StudentTest {
+
+	@Before
+	public void setUp() throws Exception {
+	}
+
+	@Test
+	public void testToString() {
+		Student student = new Student();
+		student.setId(1);
+		student.setEmail("stud1@mail.com");
+		String expected = "Student [id=1, email=stud1@mail.com]";
+		assertEquals(expected , student.toString());
+	}
+
+}
